SonarQ/requirements.txt
requests
python-dotenv
# aider-chat removed due to Python 3.13 compatibility issues
# Alternative AI coding tools that work with Python 3.13:
openai
anthropic
google-generativeai
